feat: rpc_server detached from aggregator
